Start collecting constants in module: executable names
Start collecting constants in a dedicated python module
`soundcraft.constants` by defining the basenames of all
executables there and using them everywhere.

The standalone `soundcraft.constants` module not importing
any other modules makes it possible to e.g. import this module
from the setuptools `setup.py` script in the top level directory
without drawing in any complex build time dependencies to other
modules.

Note that the actual location of the source file imported as
`soundcraft.constants` could differ depending on what the user
has set `PYTHONPATH` to or modified `sys.path` to in some startup
file.  Therefore, we verify that we have actually imported
the `soundcraft/constants.py` source file which is located below
the directory above the `setup.py`.

Importing the constants module both avoids the difficulty of
regexp parsing a programming language of which only a subset can
be parsed with regexps, and also allows us to use more of the
power of a programming language in the `soundcraft.constants`
module like using some constant values to compose others.