XXX setup: Create a shell script to run via sudo
Collect commands into a shell script which is then presented
to the user running `soundcraft_setup` for manual examination
and subsequent execution e.g. via sudo.

We avoid running sudo by ourselves to make it obvious to the
person running `soundcraft_setup` what the actual commands run
by `sudo` will be and to have them make an informed decision.

If `--sudo-script=FILENAME` is given, then the script content is
written to the given `FILENAME` which can then be executed by sudo.
Otherwise, the script is printed on stdout and can the user can
cut and paste commands from that.

The sudo script will be used to (un)install udev rules files and
remove artefacts from older installations of `soundcraft-utils`.

We will be keeping track of the actual changes which will be
made by the commands in the sudo script, so we only add a command
to the script if it will actually changes something.