setup: Add `--chroot=CHROOT` argument
When the `--chroot=CHROOT` argument is given on the `soundcraft_setup`
command line, prefix the actual (un)install location of the files with
the given `CHROOT` path.

This allows e.g. to logically install to `/usr` (including the
executables' absolute paths inside `*.desktop` and `*.service`),
but prefix the whole thing to install as non-root user into a
directory tree like `/tmp/soundcraft-utils-package-root/usr`
writable by a non-root user.

This is a required feature for many distro package build processes.