setup: check module imports before using them
Check that module imports work in `soundcraft_setup`, i.e. during
installation.  That is a better time to check than at runtime
when the user runs the CLI or the GUI.

These checks can only work with a direct installation to a
location in the filesystem; they are deactivated when installing
into a chroot directory.

Closes: https://github.com/lack/soundcraft-utils/issues/29