HACKING.md: Please make short source lines unique
If your Python code adds or touches a very short line which is also
in other places in the same source file (e.g. `try:`, `pass`,
`next`, or `continue`), please try adding a useful comment to the
line which makes the line unique enough that git's diff generator
cannot confuse this instance of the line with a completely different
instance from another part of the code.